Fetch issues straight from the Refactoring_Critical_CSS_State_Storage
