+++ projects/plugins/boost/app/assets/src/js/stores/critical-css-status.ts
@@ -1,7 +1,7 @@
 import { derived, writable } from 'svelte/store';
 import api from '../api/api';
+import { CriticalCssIssue } from './ds-critical-css-recommendations';
 import { modules } from './modules';
-import type { CriticalCssIssue } from './critical-css-recommendations';
 import type { ProviderKeyUrls, ProvidersSuccessRatio } from '../utils/generate-critical-css';
 import type { JSONObject } from '../utils/json-types';
 import type { Viewport } from '../utils/types';
